+++ sqlprojection/adaptor.go
@@ -3,7 +3,9 @@ package sqlprojection
 import (
 	"context"
 	"database/sql"
+	"sync/atomic"
 
+	"github.com/dogmatiq/cosyne"
 	"github.com/dogmatiq/dogma"
 	"github.com/dogmatiq/projectionkit/internal/identity"
 	"github.com/dogmatiq/projectionkit/internal/unboundhandler"
@@ -14,9 +16,13 @@ import (
 type adaptor struct {
 	MessageHandler
 
-	db     *sql.DB
-	key    string
-	driver Driver
+	db  *sql.DB
+	key string
+
+	m          cosyne.Mutex
+	resolved   int32 // atomic bool, fast path
+	candidates []Driver
+	selected   Driver
 }
 
 // New returns a new Dogma projection message handler by binding an SQL-specific
@@ -25,47 +31,31 @@ type adaptor struct {
 // If db is nil the returned handler will return an error whenever an operation
 // that requires the database is performed.
 //
-// If d is nil, the appropriate default driver is selected automatically if
-// possible.
+// By default an appropriate Driver implementation is chosen from the built-in
+// drivers listed in the Drivers slice.
 func New(
 	db *sql.DB,
 	h MessageHandler,
-	d Driver,
-) (dogma.ProjectionMessageHandler, error) {
+	options ...Option,
+) dogma.ProjectionMessageHandler {
 	if db == nil {
-		return unboundhandler.New(h), nil
-	}
-
-	if d == nil {
-		var err error
-		d, err = NewDriver(db)
-		if err != nil {
-			return nil, err
-		}
+		return unboundhandler.New(h)
 	}
 
 	a := &adaptor{
 		MessageHandler: h,
 		db:             db,
-		driver:         d,
 		key:            identity.Key(h),
 	}
 
-	return a, nil
-}
+	if len(options) == 0 {
+		options = []Option{
+			WithCandidateDrivers(BuiltInDrivers()...),
+		}
+	}
 
-// MustNew returns a new projection message handler that uses the given database
-// or panics if unable to do so.
-//
-// If d is nil, the appropriate default driver for db is used, if recognized.
-func MustNew(
-	db *sql.DB,
-	h MessageHandler,
-	d Driver,
-) dogma.ProjectionMessageHandler {
-	a, err := New(db, h, d)
-	if err != nil {
-		panic(err)
+	for _, opt := range options {
+		opt.applyAdaptorOption(a)
 	}
 
 	return a
@@ -78,42 +68,44 @@ func (a *adaptor) HandleEvent(
 	s dogma.ProjectionEventScope,
 	m dogma.Message,
 ) (bool, error) {
-	tx, err := a.db.BeginTx(ctx, nil)
-	if err != nil {
-		return false, err
-	}
-	defer tx.Rollback()
+	return a.withTx(ctx, func(d Driver, tx *sql.Tx) (bool, error) {
+		ok, err := d.UpdateVersion(ctx, tx, a.key, r, c, n)
+		if !ok || err != nil {
+			return ok, err
+		}
 
-	ok, err := a.driver.UpdateVersion(ctx, tx, a.key, r, c, n)
-	if !ok || err != nil {
-		return ok, err
-	}
-
-	if err := a.MessageHandler.HandleEvent(ctx, tx, s, m); err != nil {
-		return false, err
-	}
-
-	return true, tx.Commit()
+		return true, a.MessageHandler.HandleEvent(ctx, tx, s, m)
+	})
 }
 
 // ResourceVersion returns the version of the resource r.
 func (a *adaptor) ResourceVersion(ctx context.Context, r []byte) ([]byte, error) {
-	return a.driver.QueryVersion(ctx, a.db, a.key, r)
+	var v []byte
+
+	return v, a.withDriver(ctx, func(d Driver) error {
+		var err error
+		v, err = d.QueryVersion(ctx, a.db, a.key, r)
+		return err
+	})
 }
 
 // CloseResource informs the projection that the resource r will not be
 // used in any future calls to HandleEvent().
 func (a *adaptor) CloseResource(ctx context.Context, r []byte) error {
-	return a.driver.DeleteResource(ctx, a.db, a.key, r)
+	return a.withDriver(ctx, func(d Driver) error {
+		return d.DeleteResource(ctx, a.db, a.key, r)
+	})
 }
 
 // StoreResourceVersion sets the version of the resource r to v
 func (a *adaptor) StoreResourceVersion(ctx context.Context, r, v []byte) error {
-	if len(v) == 0 {
-		return a.driver.DeleteResource(ctx, a.db, a.key, r)
-	}
+	return a.withDriver(ctx, func(d Driver) error {
+		if len(v) == 0 {
+			return d.DeleteResource(ctx, a.db, a.key, r)
+		}
 
-	return a.driver.StoreVersion(ctx, a.db, a.key, r, v)
+		return d.StoreVersion(ctx, a.db, a.key, r, v)
+	})
 }
 
 // UpdateResourceVersion updates the version of the resource r to n without
@@ -123,28 +115,93 @@ func (a *adaptor) StoreResourceVersion(ctx context.Context, r, v []byte) error {
 func (a *adaptor) UpdateResourceVersion(
 	ctx context.Context,
 	r, c, n []byte,
-) (ok bool, err error) {
-	tx, err := a.db.BeginTx(ctx, nil)
-	if err != nil {
-		return false, err
-	}
-	defer tx.Rollback()
-
-	ok, err = a.driver.UpdateVersion(ctx, tx, a.key, r, c, n)
-	if !ok || err != nil {
-		return ok, err
-	}
-
-	return true, tx.Commit()
+) (bool, error) {
+	return a.withTx(ctx, func(d Driver, tx *sql.Tx) (bool, error) {
+		return d.UpdateVersion(ctx, tx, a.key, r, c, n)
+	})
 }
 
 // DeleteResource removes all information about the resource r from the
 // handler's data store.
 func (a *adaptor) DeleteResource(ctx context.Context, r []byte) error {
-	return a.driver.DeleteResource(ctx, a.db, a.key, r)
+	return a.withDriver(ctx, func(d Driver) error {
+		return d.DeleteResource(ctx, a.db, a.key, r)
+	})
 }
 
 // Compact reduces the size of the projection's data.
 func (a *adaptor) Compact(ctx context.Context, s dogma.ProjectionCompactScope) error {
 	return a.MessageHandler.Compact(ctx, a.db, s)
+}
+
+// withDriver calls fn with the driver that the adaptor should use.
+func (a *adaptor) withDriver(
+	ctx context.Context,
+	fn func(Driver) error,
+) error {
+	d, err := a.driver(ctx)
+	if err != nil {
+		return err
+	}
+
+	return fn(d)
+}
+
+// withTx calls fn with the driver that the adaptor should use.
+func (a *adaptor) withTx(
+	ctx context.Context,
+	fn func(Driver, *sql.Tx) (bool, error),
+) (bool, error) {
+	var ok bool
+
+	return ok, a.withDriver(
+		ctx,
+		func(d Driver) error {
+			tx, err := a.db.BeginTx(ctx, nil)
+			if err != nil {
+				return err
+			}
+			defer tx.Rollback()
+
+			ok, err = fn(d, tx)
+			if err != nil {
+				return err
+			}
+
+			if ok {
+				return tx.Commit()
+			}
+
+			return tx.Rollback()
+		},
+	)
+}
+
+// driver returns the driver that should be used by the adaptor.
+func (a *adaptor) driver(ctx context.Context) (Driver, error) {
+	if atomic.LoadInt32(&a.resolved) == 0 {
+		// If the resolved flag is 0 then a.selected has not been populated yet.
+		// We acquire the mutex to ensure we're the only goroutine attempting
+		// selection.
+		if err := a.m.Lock(ctx); err != nil {
+			return nil, err
+		}
+		defer a.m.Unlock()
+
+		// Ensure that no another goroutine selected the driver while we were
+		// waiting to acquire the mutex.
+		if atomic.LoadInt32(&a.resolved) == 0 {
+			// If not, it's our turn to try selection.
+			d, err := SelectDriver(ctx, a.db, a.candidates)
+			if err != nil {
+				return nil, err
+			}
+
+			a.candidates = nil
+			a.selected = d
+			atomic.StoreInt32(&a.resolved, 1)
+		}
+	}
+
+	return a.selected, nil
 }